Create ScopedRoleResourceDefinition to encapsulate common role attribute definition creation/registration.
