+++ domain-management/src/main/java/org/jboss/as/domain/management/access/ServerGroupScopedRoleResourceDefinition.java
@@ -18,13 +18,11 @@ import org.jboss.as.controller.AttributeDefinition;
 import org.jboss.as.controller.AttributeMarshaller;
 import org.jboss.as.controller.ListAttributeDefinition;
 import org.jboss.as.controller.OperationDefinition;
+import org.jboss.as.controller.OperationStepHandler;
 import org.jboss.as.controller.PathElement;
-import org.jboss.as.controller.ReloadRequiredWriteAttributeHandler;
-import org.jboss.as.controller.SimpleAttributeDefinition;
 import org.jboss.as.controller.SimpleAttributeDefinitionBuilder;
 import org.jboss.as.controller.SimpleListAttributeDefinition;
 import org.jboss.as.controller.SimpleOperationDefinitionBuilder;
-import org.jboss.as.controller.SimpleResourceDefinition;
 import org.jboss.as.controller.access.constraint.ServerGroupEffectConstraint;
 import org.jboss.as.controller.access.management.WritableAuthorizerConfiguration;
 import org.jboss.as.controller.descriptions.ModelDescriptionConstants;
@@ -41,15 +39,9 @@ import org.jboss.dmr.ModelType;
  *
  * @author Brian Stansberry (c) 2013 Red Hat Inc.
  */
-public class ServerGroupScopedRoleResourceDefinition extends SimpleResourceDefinition {
+public class ServerGroupScopedRoleResourceDefinition extends ScopedRoleResourceDefinition {
 
     public static final PathElement PATH_ELEMENT = PathElement.pathElement(SERVER_GROUP_SCOPED_ROLE);
-
-    public static final SimpleAttributeDefinition BASE_ROLE =
-            new SimpleAttributeDefinitionBuilder(ModelDescriptionConstants.BASE_ROLE, ModelType.STRING)
-            .setRestartAllServices()
-            .build();
-
 
     public static final ListAttributeDefinition SERVER_GROUPS =
             SimpleListAttributeDefinition.Builder.of(ModelDescriptionConstants.SERVER_GROUPS,
@@ -66,12 +58,12 @@ public class ServerGroupScopedRoleResourceDefinition extends SimpleResourceDefin
             .setWrapXmlList(false)
             .build();
 
-    private final ServerGroupScopedRoleAdd addHandler;
-    private final ServerGroupScopedRoleRemove removeHandler;
-    private final ServerGroupScopedRoleWriteAttributeHandler writeAttributeHandler;
+    private final OperationStepHandler addHandler;
+    private final OperationStepHandler removeHandler;
+    private final OperationStepHandler writeAttributeHandler;
 
     public ServerGroupScopedRoleResourceDefinition(WritableAuthorizerConfiguration authorizerConfiguration) {
-        super(PATH_ELEMENT, DomainManagementResolver.getResolver("core.access-control.server-group-scoped-role"));
+        super(PATH_ELEMENT, DomainManagementResolver.getResolver("core.access-control.server-group-scoped-role"), authorizerConfiguration);
 
         Map<String, ServerGroupEffectConstraint> constraintMap = new HashMap<String, ServerGroupEffectConstraint>();
         this.addHandler = new ServerGroupScopedRoleAdd(constraintMap, authorizerConfiguration);
@@ -91,7 +83,6 @@ public class ServerGroupScopedRoleResourceDefinition extends SimpleResourceDefin
     public void registerAttributes(ManagementResourceRegistration resourceRegistration) {
         super.registerAttributes(resourceRegistration);
 
-        resourceRegistration.registerReadWriteAttribute(BASE_ROLE, null, new ReloadRequiredWriteAttributeHandler(BASE_ROLE));
         resourceRegistration.registerReadWriteAttribute(SERVER_GROUPS, null, writeAttributeHandler);
     }
 }
